+++ pkg/client/client.go
@@ -1,13 +1,11 @@
 package client
 
 import (
-	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
 
 	"github.com/rocketblend/rocketblend/pkg/core/addon"
-	"github.com/rocketblend/rocketblend/pkg/core/executable"
 	"github.com/rocketblend/rocketblend/pkg/core/install"
 	"github.com/rocketblend/rocketblend/pkg/core/library"
 	"github.com/rocketblend/rocketblend/pkg/core/preference"
@@ -142,358 +140,6 @@ func (c *Client) Initialize() error {
 	return nil
 }
 
-func (c *Client) Platform() runtime.Platform {
-	return c.conf.Platform
-}
-
-func (c *Client) FindResource(key string) (*resource.Resource, error) {
-	// TODO: save out if not found.
-	return c.resource.FindByName(key)
-}
-
-func (c *Client) FindInstall(ref string) (*install.Install, error) {
-	id := c.encoder.Hash(ref)
-	install, err := c.install.FindByID(id)
-	if err != nil {
-		return nil, err
-	}
-
-	return install, nil
-}
-
-func (c *Client) FindAddon(ref string) (*addon.Addon, error) {
-	id := c.encoder.Hash(ref)
-	addon, err := c.addon.FindByID(id)
-	if err != nil {
-		return nil, err
-	}
-
-	return addon, nil
-}
-
-func (c *Client) AddInstall(path string) error {
-	build, err := c.library.FindBuildByPath(path)
-	if err != nil {
-		return err
-	}
-
-	i, _ := c.FindInstall(build.Reference)
-	if i != nil {
-		return fmt.Errorf("build already installed")
-	}
-
-	err = c.install.Create(c.newInstall(build.Reference, path, build.Packages))
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (c *Client) AddAddon(path string) error {
-	pack, err := c.library.FindPackageByPath(path)
-	if err != nil {
-		return err
-	}
-
-	a, _ := c.FindAddon(pack.Reference)
-	if a != nil {
-		return fmt.Errorf("addon already installed")
-	}
-
-	err = c.addon.Create(c.newAddon(pack.Reference, path))
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (c *Client) InstallBuild(ref string) error {
-	// Check if install already exists
-	inst, _ := c.FindInstall(ref)
-	if inst != nil {
-		return fmt.Errorf("already installed")
-	}
-
-	// Fetch build from ref
-	build, err := c.library.FetchBuild(ref)
-	if err != nil {
-		return err
-	}
-
-	if build == nil {
-		return fmt.Errorf("invalid build")
-	}
-
-	// Output directory
-	outPath := filepath.Join(c.conf.InstallationDir, ref)
-
-	// Create output directories
-	err = os.MkdirAll(outPath, os.ModePerm)
-	if err != nil {
-		return err
-	}
-
-	// build info for current platform
-	source := build.GetSourceForPlatform(c.conf.Platform)
-	if source == nil {
-		return fmt.Errorf("no source found for platform %s", c.conf.Platform)
-	}
-
-	// Download URL
-	downloadURL := source.URL
-
-	// Download file path
-	name := filepath.Base(downloadURL)
-	filePath := filepath.Join(outPath, name)
-
-	// Download file to file path
-	err = c.downloader.Download(downloadURL, filePath)
-	if err != nil {
-		return err
-	}
-
-	// Extract the archived file
-	if err := c.archiver.Extract(filePath); err != nil {
-		return err
-	}
-
-	// Markshal build
-	data, err := json.Marshal(build)
-	if err != nil {
-		return err
-	}
-
-	// Write out build.json
-	if err := os.WriteFile(filepath.Join(outPath, library.BuildFile), data, os.ModePerm); err != nil {
-		return err
-	}
-
-	// Add install to database
-	err = c.install.Create(c.newInstall(ref, outPath, build.Packages))
-	if err != nil {
-		return err
-	}
-
-	// TODO: call asynchronously
-	// Install packages
-	for _, p := range build.Packages {
-		err = c.installPackageIgnorable(p, true)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (c *Client) InstallPackage(ref string) error {
-	return c.installPackageIgnorable(ref, false)
-}
-
-func (c *Client) RemoveInstall(hash string) error {
-	return fmt.Errorf("not implemented")
-}
-
-func (c *Client) RemoveAddon(hash string) error {
-	return fmt.Errorf("not implemented")
-}
-
-func (c *Client) FindAllInstalls() ([]*install.Install, error) {
-	ints, err := c.install.FindAll()
-	if err != nil {
-		return nil, err
-	}
-
-	return ints, nil
-}
-
-func (c *Client) FindAllAddons() ([]*addon.Addon, error) {
-	addons, err := c.addon.FindAll()
-	if err != nil {
-		return nil, err
-	}
-
-	return addons, nil
-}
-
-func (c *Client) SetDefaultExecutable(ref string) error {
-	settings, err := c.preference.Find()
-	if err != nil {
-		return err
-	}
-
-	if settings == nil {
-		settings = &preference.Settings{}
-	}
-
-	settings.DefaultBuild = ref
-
-	err = c.preference.Create(settings)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (c *Client) FindDefaultExecutable() (*executable.Executable, error) {
-	settings, err := c.preference.Find()
-	if err != nil {
-		return nil, err
-	}
-
-	if settings.DefaultBuild == "" {
-		// TODO: Get latest build and set as default
-		return nil, fmt.Errorf("no default executable set")
-	}
-
-	executable, err := c.FindExecutableByBuildReference(settings.DefaultBuild)
-	if err != nil {
-		return nil, err
-	}
-
-	return executable, nil
-}
-
-// func (c *Client) FindBuildByPath(build string) (*library.Build, error) {
-// 	return c.library.FindBuildByPath(build)
+// func (c *Client) Platform() runtime.Platform {
+// 	return c.conf.Platform
 // }
-
-// func (c *Client) FetchPackageByPath(pack string) (*library.Package, error) {
-// 	return c.library.FindPackageByPath(pack)
-// }
-
-func (c *Client) FindExecutableByBuildReference(ref string) (*executable.Executable, error) {
-	if ref == "" {
-		return nil, fmt.Errorf("invalid build reference")
-	}
-
-	install, err := c.FindInstall(ref)
-	if err != nil {
-		return nil, fmt.Errorf("failed to find install: %s", err)
-	}
-
-	build, err := c.library.FindBuildByPath(install.Path)
-	if err != nil {
-		return nil, fmt.Errorf("failed to find build: %s", err)
-	}
-
-	addonMap, err := c.GetAddonMapByReferences(build.Packages)
-	if err != nil {
-		return nil, fmt.Errorf("failed to find all addons for build: %s", err)
-	}
-
-	return &executable.Executable{
-		Path:   filepath.Join(install.Path, build.GetSourceForPlatform(c.conf.Platform).Executable),
-		Addons: addonMap,
-		ARGS:   build.Args,
-	}, nil
-}
-
-func (c *Client) GetAddonMapByReferences(ref []string) (map[string]string, error) {
-	addonMap := make(map[string]string)
-	for _, a := range ref {
-		name, path, err := c.getAddonNamePathByReference(a)
-		if err != nil {
-			return nil, fmt.Errorf("failed to find package: %s", err)
-		}
-
-		addonMap[name] = path
-	}
-
-	return addonMap, nil
-}
-
-func (c *Client) getAddonNamePathByReference(ref string) (string, string, error) {
-	addon, err := c.FindAddon(ref)
-	if err != nil {
-		return "", "", fmt.Errorf("failed to find addon: %s", err)
-	}
-
-	pack, err := c.library.FindPackageByPath(addon.Path)
-	if err != nil {
-		return "", "", fmt.Errorf("failed to find package: %s", err)
-	}
-
-	return pack.Name, filepath.Join(addon.Path, pack.Source.File), nil
-}
-
-func (c *Client) installPackageIgnorable(ref string, ignore bool) error {
-	// TODO: Move downloading packages/builds into library service.
-
-	// Check if addon already exists
-	adn, _ := c.FindAddon(ref)
-	if adn != nil {
-		if !ignore {
-			return fmt.Errorf("already installed")
-		}
-		return nil
-	}
-
-	// Fetch package from ref
-	pack, err := c.library.FetchPackage(ref)
-	if err != nil {
-		return err
-	}
-
-	if pack == nil {
-		return fmt.Errorf("invalid package")
-	}
-
-	// Output directory
-	outPath := filepath.Join(c.conf.InstallationDir, ref)
-
-	// Create output directories
-	err = os.MkdirAll(outPath, os.ModePerm)
-	if err != nil {
-		return err
-	}
-
-	// download file path
-	name := filepath.Base(pack.Source.URL)
-	filePath := filepath.Join(outPath, name)
-
-	// Download file to file path
-	err = c.downloader.Download(pack.Source.URL, filePath)
-	if err != nil {
-		return err
-	}
-
-	// Markshal pack
-	data, err := json.Marshal(pack)
-	if err != nil {
-		return err
-	}
-
-	// Write out package.json
-	if err := os.WriteFile(filepath.Join(outPath, library.PackgeFile), data, os.ModePerm); err != nil {
-		return err
-	}
-
-	// Add addon to database
-	err = c.addon.Create(c.newAddon(ref, outPath))
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (c *Client) newInstall(ref string, path string, packs []string) *install.Install {
-	return &install.Install{
-		Id:       c.encoder.Hash(ref),
-		Build:    ref,
-		Path:     path,
-		Packages: packs,
-	}
-}
-
-func (c *Client) newAddon(ref string, path string) *addon.Addon {
-	return &addon.Addon{
-		Id:      c.encoder.Hash(ref),
-		Package: ref,
-		Path:    path,
-	}
-}
