+++ pkg/client/resource.go
@@ -2,7 +2,7 @@ package client
 
 import "github.com/rocketblend/rocketblend/pkg/core/resource"
 
-func NewResourceService(dir string) *resource.Service {
-	srv := resource.NewService(dir)
-	return srv
+func (c *Client) FindResource(key string) (*resource.Resource, error) {
+	// TODO: save out if not found.
+	return c.resource.FindByName(key)
 }
